+++ src/javascript/lib/core.js
@@ -13,6 +13,7 @@ import binary from './core/erlang_compat/binary';
 import unicode from './core/erlang_compat/unicode';
 import Store from './core/store';
 import math from './core/erlang_compat/math';
+import proplists from './core/erlang_compat/proplists';
 
 class Integer {}
 class Float {}
@@ -40,6 +41,7 @@ export default {
   Tuple: ErlangTypes.Tuple,
   PID: ErlangTypes.PID,
   BitString: ErlangTypes.BitString,
+  Reference: ErlangTypes.Reference,
   Patterns,
   Integer,
   Float,
@@ -56,4 +58,5 @@ export default {
   unicode,
   elixir_config,
   math,
+  proplists,
 };